+++ sperl_resolved_type.h
@@ -17,7 +17,7 @@ enum {
 };
 
 extern const char* const SPerl_RESOLVED_TYPE_C_CORE_NAMES[];
-extern int64_t const SPerl_RESOLVED_TYPE_C_CORE_SIZES[];
+extern int32_t const SPerl_RESOLVED_TYPE_C_CORE_SIZES[];
 
 extern const char* const SPerl_RESOLVED_TYPE_C_CORE_ARRAY_NAMES[];
 
@@ -35,7 +35,7 @@ _Bool SPerl_RESOLVED_TYPE_contain_sub(SPerl* sperl, SPerl_RESOLVED_TYPE* resolve
 _Bool SPerl_RESOLVED_TYPE_is_integral(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
 _Bool SPerl_RESOLVED_TYPE_is_core_type_array(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
 _Bool SPerl_RESOLVED_TYPE_is_core_type(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
-int64_t SPerl_RESOLVED_TYPE_get_array_dimention(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
+int32_t SPerl_RESOLVED_TYPE_get_array_dimention(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
 _Bool SPerl_RESOLVED_TYPE_is_calculatable_type(SPerl* sperl, SPerl_RESOLVED_TYPE* resolved_type);
 
 #endif
