+++ sperl_toke.c
@@ -70,7 +70,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
             else {
               
               // Change :: to / and add ".spvm"
-              int64_t module_path_base_length = (int64_t)strlen(package_name) + 6;
+              size_t module_path_base_length = strlen(package_name) + 6;
               char* module_path_base = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, module_path_base_length);
               const char* bufptr_orig = package_name;
               char* bufptr_to = module_path_base;
@@ -93,12 +93,12 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
               // Search module file
               char* cur_module_path = NULL;
               FILE* fh = NULL;
-              int64_t include_pathes_length = parser->include_pathes->length;
-              for (int64_t i = 0; i < include_pathes_length; i++) {
+              size_t include_pathes_length = parser->include_pathes->length;
+              for (size_t i = 0; i < include_pathes_length; i++) {
                 const char* include_path = (const char*) SPerl_ARRAY_fetch(sperl, parser->include_pathes, i);
                 
                 // File name
-                int64_t file_name_length = (int64_t)strlen(include_path) + 1 + (int64_t)strlen(module_path_base);
+                size_t file_name_length = strlen(include_path) + 1 + strlen(module_path_base);
                 cur_module_path = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, file_name_length);
                 sprintf(cur_module_path, "%s/%s", include_path, module_path_base);
                 cur_module_path[file_name_length] = '\0';
@@ -126,14 +126,14 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
               
               // Read file content
               fseek(fh, 0, SEEK_END);
-              int64_t file_size = (int64_t)ftell(fh);
+              size_t file_size = ftell(fh);
               if (file_size < 0) {
                 fprintf(stderr, "Can't read file %s at %s line %" PRId64 "\n", cur_module_path, op_use->file, op_use->line);
                 exit(1);
               }
               fseek(fh, 0, SEEK_SET);
               char* src = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, file_size);
-              if ((int64_t)fread(src, 1, file_size, fh) < file_size) {
+              if (fread(src, 1, file_size, fh) < file_size) {
                 if (op_use) {
                   fprintf(stderr, "Can't read file %s at %s line %" PRId64 "\n", cur_module_path, op_use->file, op_use->line);
                 }
@@ -418,7 +418,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
             exit(1);
           }
           
-          int64_t str_len = (int64_t)(parser->bufptr - cur_token_ptr);
+          size_t str_len = (parser->bufptr - cur_token_ptr);
           str = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, str_len);
           memcpy(str, cur_token_ptr, str_len);
           str[str_len] = '\0';
@@ -449,7 +449,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
             parser->bufptr++;
           }
           
-          int64_t str_len = (int64_t)(parser->bufptr - cur_token_ptr);
+          size_t str_len = (parser->bufptr - cur_token_ptr);
           char* var_name = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, str_len);
           memcpy(var_name, cur_token_ptr, str_len);
           var_name[str_len] = '\0';
@@ -486,7 +486,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
           }
           
           // Number literal(first is space for sign)
-          int64_t str_len = (int64_t)(parser->bufptr - cur_token_ptr);
+          size_t str_len = (parser->bufptr - cur_token_ptr);
           char* num_str = (char*) SPerl_ALLOCATOR_UTIL_safe_malloc(str_len + 2, sizeof(char));
           memcpy(num_str, cur_token_ptr, str_len);
           num_str[str_len] = '\0';
@@ -538,7 +538,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
           }
           // long
           else if (constant->code == SPerl_CONSTANT_C_CODE_LONG) {
-            int64_t num;
+            int32_t num;
             errno = 0;
             if (num_str[0] == '0' && num_str[1] == 'x') {
               num = strtol(num_str, &end, 16);
@@ -583,7 +583,7 @@ int SPerl_yylex(SPerl_YYSTYPE* yylvalp, SPerl* sperl) {
           }
           
           
-          int64_t str_len = (int64_t)(parser->bufptr - cur_token_ptr);
+          size_t str_len = (parser->bufptr - cur_token_ptr);
           char* keyword = SPerl_ALLOCATOR_PARSER_new_string(sperl, parser, str_len);
           
           memcpy(keyword, cur_token_ptr, str_len);
